Migrate fromisoformat tests to their own file
