+++ Lib/test/test_datetime.py
@@ -8,9 +8,7 @@ TESTS = 'test.datetimetester'
 
 def load_tests(loader, tests, pattern):
     try:
-        pure_tests = import_fresh_module(TESTS, fresh=[
-            'datetime', '_strptime', 'test.isoformat_helpers.isoformatter',
-            'test.isoformat_helpers.strategies'],
+        pure_tests = import_fresh_module(TESTS, fresh=['datetime', '_strptime'],
                                         blocked=['_datetime'])
         fast_tests = import_fresh_module(TESTS, fresh=['datetime',
                                                     '_datetime', '_strptime'])
